+++ src/events/changeconfiguration.ts
@@ -1,5 +1,7 @@
 import * as vscode from "vscode";
+import { settingsManager } from "../extension/settings";
 
 
 export function onDidChangeConfiguration(configEvent: vscode.ConfigurationChangeEvent){
+    settingsManager.updateSettings(configEvent);
 }