+++ src/user/usermanager.ts
@@ -16,9 +16,23 @@ class UserManager {
     private isLoggedin: boolean;
     private awaitsLogin: boolean = false;
 
+    private refreshTimer: NodeJS.Timer;
+
     public set IsLoggedin(value: boolean) {
         this.isLoggedin = value;
         this.session.IsLoggedin = value;
+
+        if (this.system.isMain)
+            SetContextValue("loggedin", value);
+
+        if (value && !this.refreshTimer) {
+            this.refreshTimer = setInterval(() => {
+                this.refreshLogin();
+            }, 10 * 60 * 1000);
+        }
+        else if (!value && this.refreshTimer) {
+            clearInterval(this.refreshTimer);
+        }
     }
 
     public get IsLoggedin() {
@@ -61,37 +75,41 @@ class UserManager {
 
     async login() {
         if (this.awaitsLogin) return false;
-        if (this.session.IsLoggedin && !this.session.didCookiesExpire()) return true;
+        if (this.IsLoggedin && !this.session.didCookiesExpire()) return true;
         this.awaitsLogin = true;
+
+        if (await this.refreshLogin(false)) {
+            logger.info("Logged in succesfully for " + this.system.name + ".(e)");
+            this.IsLoggedin = true;
+            this.awaitsLogin = false;
+            return true;
+        }
         await this.setAuth();
-        const response = await logInService.call({ host: this.system.host, port: this.system.port }, this.system.name);
+        const response = await logInService.call({ host: this.system.host, port: this.system.port }, true, { Session: false });
         this.awaitsLogin = false;
+
         if (response) {
-            logger.info("Login successful for " + this.system.name);
+            logger.info("Logged in succesfully for " + this.system.name + ".(n)");
             this.session.haveCookies(response);
             this.IsLoggedin = true;
-            if (this.system.isMain)
-                SetContextValue("loggedin", true);
-
-            if (!this.refreshTimer)
-                this.refreshTimer = setInterval(() => { this.refreshLogin() }, 5 * 60 * 1000)
-
             return true;
         }
         else {
-            logger.error("Login not successful for " + this.system.name);
+            logger.error("Log in not successful for " + this.system.name);
         }
         return false;
     }
 
-    refreshTimer: NodeJS.Timer;
+
     // extension setting changes this 
-    async refreshLogin() {
-        if(this.session.loadCookiesIfCookedIn(5)) return;
-        const response = await logInService.call({ host: this.system.host, port: this.system.port }, this.system.name);
+    async refreshLogin(useCookies = true) {
+        if (useCookies && this.session.loadCookiesIfCookedIn(10)) return true;
+        const response = await logInService.call({ host: this.system.host, port: this.system.port }, false, { Session: true });
         if (response) {
             this.session.haveCookies(response);
+            return true;
         }
+        return false;
     }
 
     async logout() {
@@ -99,15 +117,10 @@ class UserManager {
         logger.info('Log out for ' + this.system.name);
         this.session.clear();
         this.IsLoggedin = false;
-        if (this.system.isMain)
-            SetContextValue("loggedin", false);
-        if (this.refreshTimer)
-            clearInterval(this.refreshTimer);
     }
 
     async setAuth(promptPassword = true) {
-        if (this.system.password == null && promptPassword && await this.askPassword()) {
-        }
+        this.system.password == null && promptPassword && await this.askPassword();
 
         const newAuth = Buffer.from(this.system.username + ":" + (this.system.password || this.password)).toString('base64');
         this.session.auth = newAuth;